backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
langchain==0.3.13
langchain-openai==0.2.14
langchain-community==0.3.13
neo4j==5.27.0
pandas==2.2.3
httpx==0.28.1
python-multipart==0.0.20
pytest==8.3.4
openai==1.59.6